🐛 fix(User.java): change parameter type from int to double in the send() method to allow for more precise transaction amounts The parameter type of the `send()` method in the `User` class has been changed from `int` to `double` to allow for more precise transaction amounts. This change improves the accuracy of transactions by supporting decimal values for the transaction amount.

+++ app/src/main/java/rhizome/core/user/User.java
@@ -63,7 +63,7 @@ public class User {
         return Transaction.of(this.getAddress(), PDN(50));
     }
 
-    public Transaction send(User receiver, int i) {
+    public Transaction send(User receiver, double i) {
         return send(receiver, PDN(i));
     }
